moved get food items into product utils
