+++ utils/helper_func.py
@@ -1,2 +1,4 @@
+SPECIAL_CHARACTERS = "[\s\-:;]"
+
 def clean_data(data):
     return data.strip().lower()
